[LC] light classes shouldn't compute icons themselves
Instead, `KotlinIconProvider` on the Kotlin IntelliJ plugin
side has to handle them. This is how it worked and should.
The only problem is the case there the plugin returns null
from the provider. This case might be either an error one
(the plugin just has no logic for this case), or on purpose
if it has no specific icon.
But this behavior produces is ambiguity for light classes, so they
have to return `null` instead of throwing an exception.

The added test ensures that all important declarations have this
override and no redundant work is done on its call

^KT-80812 Fixed