[FIR] JvmMappedScope: add missed lazy resolve
JvmMappedScope is supposed to operate with binary versions
of classes, but in the case of Kotlin repo with have stdlib
sources instead of the library dependency.
As a result, `firKotlinClass` might represent such classes as
`kotlin/collections/Map`. Without explicit lazy resolve,
its members are in RAW_FIR phase that means that `computeJvmDescriptor`
won't work for them since it operates with `FirFunction` directly.
`TYPES` should be enough since we assume that public stdlib declarations
don't have implicit types

^KT-82772 Fixed