Fix `IrSimpleFunction.isLegacyOpenFunctionWithDefault`
Compilation of overrides of `@Composable` methods can fail when the
methods are arranged across packages in certain ways, e.g. how they are
arranged in
`ComposeCrossModuleTests.testLegacyOpenFunctionWithDefaultFalsePositives`.
This CL fixes the root cause of the failures by making
`IrSimpleFunction.isLegacyOpenFunctionWithDefault` account for the fact
that `composeMetadata` can only appear on the root of an override tree.

Test: ComposeCrossModuleTests.testLegacyOpenFunctionWithDefaultFalsePositives
Bug: 461766771