[IR][K/N] Do not add casts from Nothing in case of box/unbox
This is a common case because of IrReturn: it has type Nothing, and if
a box/unbox operation is needed (like the function returns Int), a cast
was also added here in case of genericSafeCasts=true.
This commit removes these casts as they are unreachable anyway but make
both IR and bitcode bigger and harder to read.

 #KT-83036