[K/N] Introduce FunctionDecl.directAccess encapsulating .binaryName
After the previous commit, a `GlobalDecl` in interop Indexer describes
how it can be accessed in the `-Xccall-mode direct` mode using the
`.directAccess` property.
It can also signal that the global is not accessible directly, storing
the explanation. This explanation is forwarded to users.

The same machinery is needed for functions. So this commit naturally
reuses `DirectAccess` (and some code that handles it) for
`FunctionDecl`, adding `FunctionDecl.directAccess`.

This is a refactoring. It doesn't change any observed behavior but
serves merely as a preparation for future changes.

This commit also adds unit tests for the value of the new property.
Some of them are to be changed within KT-83039.

^KT-79755