[LL] properly process local super types
`FirSupertypeResolverVisitor.resolveSpecificClassLikeSupertypes`
uses `supertypeComputationSession` to decide whether the declaration
was already resolved. If not, then it has to be processed by
the supertype transformer (the local one).
Since the lazy resolve result wasn't stored in the session,
the compiler began to transform the declaration like it was local.
As a result, we had two problems:
1. Compiled might modify already resolved declaration
2. The empty list was stored as the result of the compiler work

The last problem led to the lack of supertype scopes since the
session is used to get supertypes.

To solve these problems, `LLSupertypeComputationSessionLocalClassesAware`
has to store the lazy resolve result.
Additionally, to simplify the logic, `supertypeRefs` could return
a non-null result, so the LL part doesn't have to care about the current
use-site session. Instead, the compiler will perform this traversal.
This is safe because the compiler won't transform non-local declarations
since `LLSupertypeComputationSessionLocalClassesAware` will be ahead of
it through `supertypeRefs`

^KT-80485 Fixed