[K/N] improve devirtualization analysis RAM consumption
Merging edges was quite memory-hungry because there are a lot
of edges in the constraing graph and hashmaps need to be sparse
to be efficient.

This change eliminated the hashsets and hashmaps and replaces them
with regular arrays. This should help reduce the peak memory consumtion
for this stage of the algorithm.

^ KT-80367