[PSI] Fix 128...256 codepoint handling in 'isIdentifier()'
The 'isIdentifier()' function assumed that there might come either plain
ASCII (A-Za-z) or Unicode (あ), and for the ASCII set, it performed the
optimized check. However, the function did the separation incorrectly –
there are legitimate Unicode letters between 128 and 256 codepoints
(such as À, æ or Ö).

The fix removes the ASCII-specific check as the `isLetter()` for ASCII
is fairly well optimized anyway.

^KTIJ-36791 Fixed
^KTIJ-37677