[IR] Return `firstRelevantLineIndex` if the offset was not found in a file
In Klib backends we introduced an optimization that can pack remove
unused offsets from `lineStartOffsets` array. Unfortunately, it
can happen that inner element can have its offsets out of bounds for
the outer element. This change ensures that even if this happens,
then we return an offset inside given file.

#KT-80814