[Tests] Fix actual debugger stepping test data generator
In case there are more actual lines then expected lines, the overflowing
lines would be put at the end of the file. Now they are put at the end
of the relevant backend's section.

This fix ensures that a correct test data is generated, even if the
"// EXPECTATIONS XXX" section is initially empty.