Fix `StabilityInferencer`'s handling of `internal` types
In the past, if an `internal` type `T` was inferred to be stable when it
was compiled in a compilation unit `A`, we would infer it to be stable
when compiling a compilation unit `B` that uses the type. This caused
failures in situations in which `T` was later made unstable by an edit and
`A` was incrementally recompiled, but `B` was not. This change eliminates
the possibility of such failures.

Test: ClassStabilityTransformTests.testInferredStabilityOfInternalTypesFromOtherCompilationUnits
Fixes: 427530633