kotlinp: Fix missing JVM metadata extensions
Previously the contents of this service implementation file were merged
with the one in kotlin-metadata-jvm, so JvmMetadataExtensions was there.
Looks like after update to Shadow 9.1, this no longer happens, so we
must manually provide the implementation in the resource file.

This doesn't affect any user-visible behavior because at the moment,
kotlinp is an internal utility.