Reflection: do not fail on error type parameters in new implementation
When a local class uses a type parameter declared in the enclosing
method, kotlin-reflect fails to load that type parameter. The old
implementation (before KT-76521) worked incorrectly too, however it did
not throw exception, but created a "error type" instead, which is wrong
in many ways, but at least it didn't fail.

The correct fix of the underlying issue (KT-47030) would be to collect
type parameters used in the class signature from enclosing methods and
their containing classes, but right now the goal is to make behavior as
close to the old (K1-based) reflection implementation as possible, so we
also create an error type here.

RCA: see above.

 #KT-83361 Fixed

(cherry picked from commit 484c869461e85ee5739fbb28922f6710ac0e5212)