[K/N] Stub for inaccessible fake overrides from ObjC
Imagine there is a final class with a fake overridden method from an internal interface which gets exported to Obj-C.
The Obj-C export machinery tries to create an adapter for this method but this might lead to a linkage error when
the signatures of the fake overridden method and the base method do not coincide. This is because a bridge would be
needed for such a call. But this bridge is not needed in Kotlin code (since this fake overridden method is from a final
class, it cannot be overridden and the codegen just shortcuts to the base method adding all coercions needed). Thus a
linkage error. It's also worth noting that even though this fake overridden gets exported, it cannot be called from the
Obj-C side (all fake overrides get filtered out).

This commit fixes the above problem by simply throwing a error from the adapter (since it's cannot be called from legitimate
code anyway).

 #KT-82160 Fixed