[K/N] Experiment with enabling CMS GC by default
CMS is available as opt-in feature since 2.0.20, and thanks to user's
feedback, we fixed several critical and major problems.
It is also enabled for Swift export by default, and some production
applications in the wild.

The next step is to roll out during 2.3.20-Beta1 and Beta2 releases and
gather additional feedback.

^KT-82635 Fixed.