[Daemon] Tolerate the case when `compilationResults` are not passed to compilation
The introduction of metrics collection for non-incremental compilations leads to some test failures with NPE. That's because of `!!` usage on `compilationResults`. The current solution is to allow passing `null` and ignore metrics collection if `compilationResults` is null. Though, it should be revised later and maybe the interface should be adjusted.
For example, currently there's a hidden requirement that for incremental compilation `compilationResults` must always be non-null, which is not nice.

Also, flush the metrics in a `finally` block