[FIR2IR] Build IR constants based on the result of FirExpressionEvaluator
Previously the FIR2IR lowering ignored the results of the frontend
interpreter rebuilt the whole FIR expression in IR and then folded them
with the backend IRInterpreter. Now for the nodes that are already
evaluated in the frontend we only insert constants into the IR.
This change will make it easier to phase the IRInterpreter out of the
compilation pipeline.

#KT-84561