[JVM] Correctly handle @JvmStatic references in Upgrade lowering
In cea6418a a special handling was added for properties, which
has JvmStatic annotation in objects. It's required to not pass
dispatch receiver to such functions, as it was already
removed in JvmStaticInObjectLowering, but still capture it
to the reference for correct reflection.

This commit adds similar supprot for function references,
which is needed to move upgrade before FunctionReferenceLowering

^KT-74383

Co-authored-by: Evgeniy.Zhelenskiy <Evgeniy.Zhelenskiy@jetbrains.com>