[Native] Generate minimalistic `NativeConfigurationKeys`
Generate it with the minimal set of keys which are required
only at the first Kotlin/Native compilation stage.

The new `NativeConfigurationKeys` class is located in a separate
module :native:native.config. Unlike `KonanConfigurationKeys`,
it is decoupled from the Kotlin/Native backend and can be
used without any dependencies on the backend modules.

^KT-78727