Fix UuidTest.testV7UuidGenerationForNonMonotonicClock flakiness
Enable the test back (it was ignored due to KTI-2674), and modify it to avoid rare failures.
The test was using an assumption that a randomly generated part of the Uuid should always differ for two consequently generated Uuids, but since that part (rand_a) is quite small (only 12 bits), the chance of a collision is relatively high (1 / 4096).
The test was rewritten to check the same property more reliably.
A side note: the implementation does not require a fix, because its behavior does not contradict RFC's recommendations: we generate a new counter when updating a timestamp, a new randomly generated value may collide with a previous one - things happen (with a low probability).

See KTI-2674 for details.

Merge-request: KT-MR-24918
Merged-by: Filipp Zhinkin <filipp.zhinkin@jetbrains.com>