KLIBs: Choose the proper module for IR deserialization with built-ins
The `depends=` property is not written to manifests of JS/Wasm Klibs
since the commit 255b2e2ebe4e33f3702f6cfa77f8157614d743f4. But it's
still used in the IR linker to recognize the "stdlib" module descriptor
that is expected to hold the built-ins: The "stdlib" module is
the module that does not depend on any other module.

So each Kotlin/JS and Kotlin/Wasm Klib library has no dependencies
written in the manifest now. And as a result, a module descriptor
created for each such library does not depend on others, and therefore
can be recognized as a descriptor with the built-ins.

This has an impact on the deserialization of IR: For regular modules
we create regular `IrModuleDeserializer`s. And only for the module
with the built-ins we create a special instance of
`IrModuleDeserializerWithBuiltIns` that is capable of providing
additional symbols such as function interfaces, etc. Having the wrong
module chosen for `IrModuleDeserializerWithBuiltIns` may result in
ineffective or incorrect deserialization.

With this commit we are fixing this by changing the way how a module
with the built-ins is recognized.

^KT-60874