[AA] Don't use compiler extension API in standalone `registerProjectExtensionPoint`
This method allows registering new extension points services, but
previously it used the `ProjectExtensionDescriptor` which is an interface
for compiler extension points. But registering new extension points for
the *compiler* is meaningless, as compiler can work only with extension
points defined by itself.
So the method was replaced with a more general one which doesn't depend
on the compiler extensions, but allows to register a general
intellij-based extension point.

^KT-83341