[Analysis API] add deprecated-hidden status for unavailable methods from JDK 21
Kotlin produces some methods (like getFirst/getLast) for compatibility
with old JDKs, and it was fixed in KT-77143, but with >=21 version
the hidden flag is still required.
The same problem is applicable for java.lang.Enum
members such as `getDeclaringClass` and `finalize`.
All such declarations are mentioned in
`JvmBuiltInsSignatures#HIDDEN_METHOD_SIGNATURES`

^KT-78550 Fixed