[K/N][IR] Fixed a bug in the types compuation pass
The bug was that the invariant was not being held: the main visitor takes a read-only
BitSet and returns a read-only BitSet. However, visitGetValue used to return a BitSet directly
from a mutable map, which might have been changed later, breaking the invariant.
The fix is to simply copy the BitSet before returning it.

 #KT-83324 Fixed