KLIB Resolver: Add test for compatibility with external legacy tools
There is at least one known client of the KLIB resolver API:
The Gradle plugin of the kotlinx-benchmarks project.

This Gradle plugin should provide binary compatibility with a wide
range of compiler versions, starting from 2.0.20. Which means
that we can't drop the KLIB resolver when it's no more used
in the compiler. Instead, we should keep it in the compiler marked
as @Deprecated. And at the same time we have to set up some tests
that would help to control that the KLIB resolver still works correctly.

With this commit, we are adding such tests.

^KT-82882
^KT-83328