KLIB Resolver: Workaround for kotlinx-benchmarks Gradle plugin
The `depends=` property is not written to manifests of JS/Wasm Klibs
since the commit 255b2e2ebe4e33f3702f6cfa77f8157614d743f4.

This has no impact on Kotlin/JS and Kotlin/Wasm compilers, since
they don't use the legacy KLIB resolver API that had some logic
based on this property.

However, it turned out that there is one legacy external client
of the KLIB resolver that still relies on this property:
The Gradle plugin of the kotlinx-benchmarks project. And this client
is affected by this change: Some libraries are not resolved; the plugin
is not able to correctly generate the code for JS/Wasm benchmarks.

In this commit we are introducing a workaround specifically for
legacy external clients such as the kotlinx-benchmarks plugin.
The workaround allows eagerly resolving all the supplied dependencies
without trying to filter those that are not mentioned in the `depends=`.

At the same time, we preserve the existing behavior of the latest
inner KLIB resolver client: Kotlin/Native compiler.

^KT-82882 Fixed