Constraint solver: introduce 'isFromFlexiblePosition' constraint flag
In this commit, we introduce 'isFromFlexiblePosition' flag for
constraints and constraint context. A constraint has this flag
in case its initial lower constraint has a flexible type variable.
This flag changes the incorporation rules: with A? <: T <: B and
A? <: T having 'isFromFlexiblePosition' we now infer
A? <: B! instead of A? <: B as before.

This commit also reverts the fix of KT-78621 (commit 8d96a64),
as the logic of simplifying Foo? <: T! to (Foo!! .. Foo?) <: T
is now superseded with 'isFromFlexibleConstraint' logic.
So Foo? <: T! now is simplified to Foo? <: T
as it was before KT-78621 fix.

These changes fix back KT-81988 and keep KT-78621 working
but break some other issues.

#KT-81988 Fixed
KT-78621 kept fixed
KT-67825 kept fixed
KT-82201 in progress
#KT-79864 Submitted
#KT-78783 Submitted