Fix mergeReleaseComposeMapping task failure when build cache is enabled
`@OutputFiles` assumes that the provided file is not a directory which is not the case for the `outputDir` property. Changing it to `@OutputDirectory` fixes the issue.

The task was also failing when the files from previous runs were already present in the file system, this change fixes that by overwriting old files on copy.

Test: ComposeIT