[K/N] further improvements in DevirtualizationAnalysis
DevirtualizationAnalysis builds a huge graph and analyses it. All the
stages except the analysis are linear in time and space. The analysis
itself is quadratic both in time and space, albeit with a good constant.

This means that making the graph smaller improves both the execution time
and the memory consumption of the analysis tremendously.

It turns out that the graph has a lot of similar structures. Merging the
nodes that are equivalent in terms of the analysis result cuts the number
of nodes to analyze by a factor of 5 (on kotlinconfapp). Removing the nodes
that don't affect the query results reduces the number of remaining nodes
by a factor of 20 (on kotlinconfapp).

All of that cuts the execution time of analysis down to 60ms and memory
consumption to a few MB's.

Of course milage may vary for a different project, but I believe these
optimizations should translate well to them as well.