[Analysis API] introduce KaImplicitInvokeCall
This is a special marker function call implementation intended to
replace the use of "KaSimpleFunctionCall#isImplicitInvoke".
On the surface, it has no relation to `KaSimpleFunctionCall`,
but the implementation has KaSimpleFunctionCall as a supertype to
preserve semantic compatibility. This means that `KaSimpleFunctionCall`
(with `isImplicitInvoke == true`) and `KaImplicitInvokeCall` can replace
each other

^KT-66039