FE: Refine rules for enhancement of covariant overrides via JSpecify
After the fix for KT-78541 (0eb1dc7d9aab408be867f4cde3c95b00ef8566de),
we started ignoring contradictions between the overridden and the
override even for the head type (top-level type). Though as far
as I can get from the description, the problematic part
was only relevant to the type arguments.

But it's obviously incorrect to top level types, especially when the
overridden is from Kotlin.

And generally, I'd call it controversial to ignore an overridden's
nullability, but it's a different story (see KT-83384).

RCA: In my opinion, the root of the problem is the lack of test data
and not a single user-project using JSpecify. The regression scenario
looks very basic, and it's unfortunate we didn't have something like
that to fail beforehand (KT-83385 to follow)

^KT-83314 Fixed
^KT-83384 Related